requirements.txt
cffi~=1.16.0
defusedxml~=0.7.1
typing_extensions~=4.9.0
pillow~=10.2.0
ipython~=8.20.0
jupyter~=1.0.0
style~=1.1.0
pip~=24.0
attrs~=23.2.0
tornado~=6.4
Jinja2~=3.1.4
pytz~=2023.4
PyYAML~=6.0.1
Django~=5.0.3
asgiref~=3.8.1
sqlparse~=0.4.4
psycopg2~=2.9.9
colorama~=0.4.6
numpy~=1.26.3
requests~=2.31.0
djangorestframework~=3.14.0
packaging~=23.2
inflection~=0.5.1
uritemplate~=4.1.1
referencing~=0.35.1
jsonschema~=4.22.0
idna~=3.6
webcolors~=1.13
jsonpointer~=2.4
isoduration~=20.11.0
fqdn~=1.5.1
platformdirs~=4.2.1
Pygments~=2.17.2